requirements.txt
Django
django-allauth
django-crispy-forms
django-debug-toolbar
whitenoise